# docs/swagger.yaml
basePath: /api
definitions:
  router.CreateDictEntityRequest:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  router.DictEntityResponse:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  router.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  router.ListDictEntitiesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/router.DictEntityResponse'
        type: array
    type: object
...
  router.UpdateDictEntityRequest:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Incident management system API
  title: Fixik API
  version: "1.0"
paths:
  /impacts:
    get:
      consumes:
      - application/json
      description: Get all impacts in dictionary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListDictEntitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List impacts
      tags:
      - impacts
    post:
      consumes:
      - application/json
      description: Create new impact dictionary entry
      parameters:
      - description: Impact data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create impact
      tags:
      - impacts
  /impacts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete impact dictionary entry by ID
      parameters:
      - description: Impact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete impact
      tags:
      - impacts
    get:
      consumes:
      - application/json
      description: Get impact dictionary entry by ID
      parameters:
      - description: Impact ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Get impact by ID
      tags:
      - impacts
    put:
      consumes:
      - application/json
      description: Update impact dictionary entry by ID
      parameters:
      - description: Impact ID
        in: path
        name: id
        required: true
        type: integer
      - description: Impact data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.UpdateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Update impact
      tags:
      - impacts
  /priorities:
    get:
      consumes:
      - application/json
      description: Get all priorities in dictionary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListDictEntitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List priorities
      tags:
      - priorities
    post:
      consumes:
      - application/json
      description: Create new status dictionary entry
      parameters:
      - description: Priority data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create status
      tags:
      - priorities
  /priorities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete status dictionary entry by ID
      parameters:
      - description: Priority ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete status
      tags:
      - priorities
    get:
      consumes:
      - application/json
      description: Get status dictionary entry by ID
      parameters:
      - description: Priority ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Get status by ID
      tags:
      - priorities
    put:
      consumes:
      - application/json
      description: Update status dictionary entry by ID
      parameters:
      - description: Priority ID
        in: path
        name: id
        required: true
        type: integer
      - description: Priority data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.UpdateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Update status
      tags:
      - priorities
  /roles:
    get:
      consumes:
      - application/json
      description: Get all roles in dictionary
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ListDictEntitiesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create new role dictionary entry
      parameters:
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.CreateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Create role
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete role dictionary entry by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Delete role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Get role dictionary entry by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Get role by ID
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update role dictionary entry by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Role data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/router.UpdateDictEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.DictEntityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/router.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/router.ErrorResponse'
      summary: Update role
      tags:
      - roles
...
schemes:
- http
swagger: "2.0"